qldb/requirements.txt
-e .
boto3==1.17.72
aws-cdk.core==1.104.0
aws-cdk.aws-qldb==1.104.0
